rest_example/requirements.txt
adium-theme-ubuntu==0.3.4
Django==1.11
django-debug-toolbar==1.8
django-pdb==0.6.1
djangorestframework==3.6.3
f.lux-indicator-applet==1.1.11rc0
gyp==0.1
pexpect==4.0.1
ptyprocess==0.5
pygobject==3.20.0
pytz==2014.10
pyxdg==0.25
sqlparse==0.2.3
unity-lens-photos==1.0
virtualenv==15.1.0
